backend/requirements.txt
Django
djangorestframework
django-cors-headers
celery
redis
psycopg2-binary
django-storages[azure]
django-storages[google]
requests

# dj-rest-auth and django-allauth for auth
dj-rest-auth[with_social]
djangorestframework-simplejwt
django-allauth

# for handling audio files and transcription
openai
pydub